feat: небольшие исправления, перебор для всех возможных ЦВЗ с тестом внедрения и извлечения осталась задача улучшения хэш функции или/и подбора констант

+++ DigitalWaterMarkApp/WaterMark/WaterMarkItem.cs
@@ -30,7 +30,7 @@ namespace DigitalWaterMarkApp
         }
 
         public WaterMarkItem(
-            byte Value,
+            int Value,
             int RowIdx,
             int ColIdx
         ) {
